+++ include/ws_client/WebSocketClientAsync.hpp
@@ -56,8 +56,6 @@ using namespace std::chrono_literals;
  *                       - `CoroioSocket`, for non-blocking async I/O based on `coroio`
  *                       - `AsioSocket`, for non-blocking async I/O based on `asio`
  * 
- * @tparam TCancelSlot   Cancellation slot/token type, e.g. asio::cancellation_slot for ASIO.
- * 
  * @tparam TMaskKeyGen   Mask key generator type. Must implement contract `HasMaskKeyOperator`.
  * 
  *                       Library implementations:
@@ -68,15 +66,13 @@ template <
     template <typename...> typename TTask,
     typename TLogger,
     typename TSocket,
-    typename TCancelSlot = typename std::remove_cvref_t<TSocket>::cancellation_slot_type,
     typename TMaskKeyGen = DefaultMaskKeyGen>
-    requires HasSocketOperationsAsync<TSocket, TTask, TCancelSlot> &&
-             HasMaskKeyOperator<TMaskKeyGen>
+    requires HasSocketOperationsAsync<TSocket, TTask> && HasMaskKeyOperator<TMaskKeyGen>
 class WebSocketClientAsync
 {
     bool closed_ = true;
 
-    BufferedSocketAsync<TSocket, TTask, TCancelSlot> socket_;
+    BufferedSocketAsync<TSocket, TTask> socket_;
     TLogger* logger_;
 
     // mask key generator
@@ -108,7 +104,7 @@ public:
     }
 
     explicit WebSocketClientAsync(TLogger* logger, TSocket&& socket) noexcept
-        : socket_(BufferedSocketAsync<TSocket, TTask, TCancelSlot>(std::move(socket))),
+        : socket_(BufferedSocketAsync<TSocket, TTask>(std::move(socket))),
           logger_(logger),
           mask_key_gen_(DefaultMaskKeyGen())
     {
@@ -153,7 +149,7 @@ public:
     /**
      * Returns the underlying socket object.
      */
-    [[nodiscard]] inline BufferedSocketAsync<TSocket, TTask, TCancelSlot>& underlying() noexcept
+    [[nodiscard]] inline BufferedSocketAsync<TSocket, TTask>& underlying() noexcept
     {
         return this->socket_;
     }
@@ -193,9 +189,7 @@ public:
      * User needs to ensure this method is called only once.
      */
     [[nodiscard]] TTask<std::expected<void, WSError>> handshake(
-        Handshake<TLogger>& handshake,
-        std::chrono::milliseconds timeout_ms = 5s,
-        TCancelSlot cancel = {}
+        Handshake<TLogger>& handshake, std::chrono::milliseconds timeout_ms = 5s
     ) noexcept
     {
         if (!this->closed_)
@@ -208,16 +202,16 @@ public:
         std::span<byte> req_data = std::span(
             reinterpret_cast<byte*>(req_str.data()), req_str.size()
         );
-        WS_CO_TRYV(co_await this->socket_.write(req_data, timeout, cancel));
+        WS_CO_TRYV(co_await this->socket_.write(req_data, timeout));
 
         // read HTTP response
         WS_CO_TRY(headers_buffer, Buffer::create(1024, 1024 * 1024)); // 1 KB to 1 MB
         byte delim[4] = {byte{'\r'}, byte{'\n'}, byte{'\r'}, byte{'\n'}};
         std::span<byte> delim_span = std::span(delim);
-        WS_CO_TRYV(co_await this->socket_.read_until(*headers_buffer, delim_span, timeout, cancel));
+        WS_CO_TRYV(co_await this->socket_.read_until(*headers_buffer, delim_span, timeout));
 
         // read and discard header terminator bytes \r\n\r\n
-        WS_CO_TRYV(co_await this->socket_.read_exact(delim_span, timeout, cancel));
+        WS_CO_TRYV(co_await this->socket_.read_exact(delim_span, timeout));
 
         // process HTTP response
         WS_CO_TRYV(handshake.process_response(string_from_bytes(headers_buffer->data())));
@@ -257,9 +251,7 @@ public:
      */
     template <HasBufferOperations TBuffer>
     [[nodiscard]] TTask<std::variant<Message, PingFrame, PongFrame, CloseFrame, WSError>>
-    read_message(
-        TBuffer& buffer, std::chrono::milliseconds timeout_ms, TCancelSlot cancel = {}
-    ) noexcept
+    read_message(TBuffer& buffer, std::chrono::milliseconds timeout_ms) noexcept
     {
         if (this->closed_)
             co_return WSError(WSErrorCode::connection_closed, "Connection in closed state.");
@@ -269,7 +261,7 @@ public:
         while (!timeout.is_expired())
         {
             // read next frame w/o payload
-            WS_CO_TRY_RAW(frame_res, co_await this->read_frame(timeout, cancel));
+            WS_CO_TRY_RAW(frame_res, co_await this->read_frame(timeout));
             Frame& frame = *frame_res;
 
             // check reserved opcodes
@@ -285,7 +277,7 @@ public:
             // handle control frames
             if (frame.header.is_control())
             {
-                auto res = co_await this->handle_control_frame(frame, timeout, cancel);
+                auto res = co_await this->handle_control_frame(frame, timeout);
 
                 // convert to outer variant
                 co_return std::visit(
@@ -402,18 +394,14 @@ public:
                         )
                     );
                     WS_CO_TRYV_RAW(
-                        co_await this->socket_.read_exact(
-                            *frame_data_compressed_res, timeout, cancel
-                        )
+                        co_await this->socket_.read_exact(*frame_data_compressed_res, timeout)
                     );
                 }
                 else
                 {
                     // read payload into message buffer
                     WS_CO_TRY_RAW(frame_data_res, buffer.append(frame.payload_size));
-                    WS_CO_TRYV_RAW(
-                        co_await this->socket_.read_exact(*frame_data_res, timeout, cancel)
-                    );
+                    WS_CO_TRYV_RAW(co_await this->socket_.read_exact(*frame_data_res, timeout));
                 }
             }
 
@@ -538,7 +526,7 @@ public:
      * NOTE: The message is always sent as a single frame, fragmentation is currently not supported.
      */
     [[nodiscard]] TTask<std::expected<void, WSError>> send_message(
-        const Message& msg, SendOptions options = {}, TCancelSlot cancel = {}
+        const Message& msg, SendOptions options = {}
     ) noexcept
     {
         if (this->closed_)
@@ -588,7 +576,7 @@ public:
         frame.set_payload_size(payload.size());
 
         Timeout timeout(options.timeout);
-        WS_CO_TRYV(co_await this->write_frame(frame, payload, timeout, cancel));
+        WS_CO_TRYV(co_await this->write_frame(frame, payload, timeout));
 
         co_return std::expected<void, WSError>{};
     }
@@ -597,7 +585,7 @@ public:
      * Sends a PONG frame to the server in response to a PING frame.
      */
     [[nodiscard]] TTask<std::expected<void, WSError>> send_pong_frame(
-        std::span<byte> payload, std::chrono::milliseconds timeout_ms = 5s, TCancelSlot cancel = {}
+        std::span<byte> payload, std::chrono::milliseconds timeout_ms = 5s
     ) noexcept
     {
         if (this->closed_)
@@ -615,7 +603,7 @@ public:
         frame.mask_key = this->mask_key_gen_();
 
         Timeout timeout(timeout_ms);
-        WS_CO_TRYV(co_await this->write_frame(frame, payload, timeout, cancel));
+        WS_CO_TRYV(co_await this->write_frame(frame, payload, timeout));
 
         co_return std::expected<void, WSError>{};
     }
@@ -647,7 +635,7 @@ public:
      * - https://www.rfc-editor.org/rfc/rfc6455#section-7.1.7
      */
     [[nodiscard]] inline TTask<std::expected<void, WSError>> close(
-        close_code code, std::chrono::milliseconds timeout_ms = 5s, TCancelSlot cancel = {}
+        close_code code, std::chrono::milliseconds timeout_ms = 5s
     ) noexcept
     {
         if (this->closed_)
@@ -673,7 +661,7 @@ public:
             );
 
             // send close frame
-            auto res = co_await this->send_close_frame(code, timeout, cancel);
+            auto res = co_await this->send_close_frame(code, timeout);
             if (!res.has_value())
             {
 #if WS_CLIENT_LOG_SEND_FRAME > 0
@@ -706,11 +694,11 @@ public:
         // This call is required to ensure all ASIO operations are cancelled
         // before closing the socket, even if the server has already closed the connection
         // in the non-graceful case.
-        co_await this->socket_.underlying().shutdown(fail_conn, timeout); // TODO: Cancel
+        co_await this->socket_.underlying().shutdown(fail_conn, timeout);
 
         // close underlying socket connection
         {
-            auto res = co_await this->socket_.underlying().close(fail_conn); // TODO: Cancel
+            auto res = co_await this->socket_.underlying().close(fail_conn);
             if (!res.has_value())
             {
 #if WS_CLIENT_LOG_TCP > 0
@@ -726,15 +714,13 @@ public:
     }
 
 private:
-    [[nodiscard]] TTask<std::expected<Frame, WSError>> read_frame(
-        Timeout<>& timeout, TCancelSlot& cancel
-    )
+    [[nodiscard]] TTask<std::expected<Frame, WSError>> read_frame(Timeout<>& timeout)
     {
         Frame frame;
 
         // read frame header (2 bytes)
         byte tmp1[2];
-        WS_CO_TRYV(co_await this->socket_.read_exact(std::span<byte>(tmp1, 2), timeout, cancel));
+        WS_CO_TRYV(co_await this->socket_.read_exact(std::span<byte>(tmp1, 2), timeout));
 
         frame.header.b0 = tmp1[0];
         frame.header.b1 = tmp1[1];
@@ -758,9 +744,7 @@ private:
             uint16_t tmp2;
             WS_CO_TRYV(
                 co_await this->socket_.read_exact(
-                    std::span<byte>(reinterpret_cast<byte*>(&tmp2), sizeof(uint16_t)),
-                    timeout,
-                    cancel
+                    std::span<byte>(reinterpret_cast<byte*>(&tmp2), sizeof(uint16_t)), timeout
                 )
             );
             payload_size = network_to_host(tmp2);
@@ -771,9 +755,7 @@ private:
             uint64_t tmp3;
             WS_CO_TRYV(
                 co_await this->socket_.read_exact(
-                    std::span<byte>(reinterpret_cast<byte*>(&tmp3), sizeof(uint64_t)),
-                    timeout,
-                    cancel
+                    std::span<byte>(reinterpret_cast<byte*>(&tmp3), sizeof(uint64_t)), timeout
                 )
             );
             payload_size = network_to_host(tmp3);
@@ -811,7 +793,7 @@ private:
     }
 
     [[nodiscard]] TTask<std::expected<void, WSError>> write_frame(
-        Frame& frame, std::span<byte> payload, Timeout<>& timeout, TCancelSlot& cancel
+        Frame& frame, std::span<byte> payload, Timeout<>& timeout
     ) noexcept
     {
 #if WS_CLIENT_LOG_SEND_FRAME > 0
@@ -867,12 +849,12 @@ private:
         frame.mask_key.mask(payload);
 
         // write frame header
-        WS_CO_TRYV(co_await this->socket_.write(write_buffer_.subspan(0, offset), timeout, cancel));
+        WS_CO_TRYV(co_await this->socket_.write(write_buffer_.subspan(0, offset), timeout));
 
         // write frame payload
         if (frame.payload_size > 0)
         {
-            WS_CO_TRYV(co_await this->socket_.write(payload, timeout, cancel));
+            WS_CO_TRYV(co_await this->socket_.write(payload, timeout));
             offset += frame.payload_size;
         }
 
@@ -880,7 +862,7 @@ private:
     }
 
     [[nodiscard]] TTask<std::expected<void, WSError>> send_close_frame(
-        close_code code, Timeout<>& timeout, TCancelSlot& cancel
+        close_code code, Timeout<>& timeout
     ) noexcept
     {
         Frame frame;
@@ -916,7 +898,7 @@ private:
 
         frame.set_payload_size(payload.size());
 
-        WS_CO_TRY(ret, co_await this->write_frame(frame, payload, timeout, cancel));
+        WS_CO_TRY(ret, co_await this->write_frame(frame, payload, timeout));
 
 #if WS_CLIENT_LOG_SEND_FRAME > 0
         logger_->template log<LogLevel::D, LogTopic::SendFrame>("Close frame sent");
@@ -926,7 +908,7 @@ private:
     }
 
     [[nodiscard]] TTask<std::variant<PingFrame, PongFrame, CloseFrame, WSError>>
-    handle_control_frame(Frame& frame, Timeout<>& timeout, TCancelSlot& cancel) noexcept
+    handle_control_frame(Frame& frame, Timeout<>& timeout) noexcept
     {
         if (!frame.header.is_final())
         {
@@ -986,9 +968,7 @@ private:
                     }
 
                     WS_CO_TRYV_RAW(
-                        co_await this->socket_.read_exact(
-                            close_frame.payload_bytes(), timeout, cancel
-                        )
+                        co_await this->socket_.read_exact(close_frame.payload_bytes(), timeout)
                     );
 
                     // check close code if provided
@@ -1029,9 +1009,7 @@ private:
                 if (frame.payload_size > 0)
                 {
                     WS_CO_TRYV_RAW(
-                        co_await this->socket_.read_exact(
-                            ping_frame.payload_bytes(), timeout, cancel
-                        )
+                        co_await this->socket_.read_exact(ping_frame.payload_bytes(), timeout)
                     );
                 }
 
@@ -1046,9 +1024,7 @@ private:
                 if (frame.payload_size > 0)
                 {
                     WS_CO_TRYV_RAW(
-                        co_await this->socket_.read_exact(
-                            pong_frame.payload_bytes(), timeout, cancel
-                        )
+                        co_await this->socket_.read_exact(pong_frame.payload_bytes(), timeout)
                     );
                 }